tests, pep8, shuffle functions and tests around
